Added a View interface which is designed to encapsulate the ASCII and GUI views of the application. It defines methods to set the answer field, get the user input and add observers. Also added the GuiViewController class, which implements the View interface which defines the behaviour for the GUI application.
I also modified the Calculator.fxml so that each element we want to interact with has IDs which we can reference. The GUIViewController class uses these for its methods to read and write to the GUI view.

Currently, the GUIViewController has an additional method called submitExpression which will evaluate a given expression and set the output string to be the result. At the moment, all that happens when it is pressed is that a stub value is displayed.